cogvlm2/requirements.txt
transformers
torch
torchvision
Pillow
accelerate
sentencepiece
